+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "test": "grunt test",
     "build": "babel src --ignore test.js --out-dir lib",
-    "prepublish": "in-publish && npm run build"
+    "prepublish": "in-publish && npm run build",
+    "test-babel": "./node_modules/mocha/bin/mocha $(find tests -name '*-test.js') --compilers js:babel-core/register"
   },
   "keywords": [
     "FlightStats",
