+++ src/index.js
@@ -1,3 +1,5 @@
 import client from './client';
+import watcher from './watcher';
+import forwarder from './forwarder';
 
 export default client;
